自动测试埋点/ios_automation/requirements.txt
Appium-Python-Client>=4.0.0,<5.0.0
PyYAML>=6.0.0,<7.0.0
requests>=2.31.0,<3.0.0
